Updates from PR review 5
