+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/manage/actions.ts
@@ -9,8 +9,7 @@ import {
 import { AuthenticatedAction } from "@lib/actions";
 import { ServerActionError } from "@lib/types/form-builder-types";
 import { logEvent } from "@lib/auditLogs";
-import { updateNotificationsSetting } from "@lib/templates";
-import { removeMarker } from "app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/notifications";
+import { removeMarker, updateNotificationsSetting } from "@lib/notifications";
 import { NotificationsInterval } from "@gcforms/types";
 
 // Public facing functions - they can be used by anyone who finds the associated server action identifer
