+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/actions.ts
@@ -13,7 +13,7 @@ import { verifyHCaptchaToken } from "@clientComponents/globals/FormCaptcha/actio
 import { checkOne } from "@lib/cache/flags";
 import { FeatureFlags } from "@lib/cache/types";
 import { validateResponses } from "@lib/validation/validation";
-import { sendNotification } from "./lib/notifications";
+import { sendNotification } from "@lib/notifications";
 
 //  Removed once hCaptcha is running in blockable mode https://github.com/cds-snc/platform-forms-client/issues/5401
 const CAPTCHA_BLOCKABLE_MODE = false;
